+++ projects/ng-toggle/package.json
@@ -1,6 +1,6 @@
 {
   "name": "ng-toggle-button",
-  "version": "0.2.3-beta.0",
+  "version": "0.2.3",
   "description": "Angular toggle button switch",
   "author": "vicmans",
   "license": "MIT",
